improve marked trail templates auto generation

+++ templates/marked_trails/marked_trails.py
@@ -382,4 +382,5 @@ if __name__ == '__main__':
          oam_mapnik_file
     )
 
-    shutil.copy2(f'../A3-{parameters.ORIENTATION.lower()}.svg', f'A3-{parameters.ORIENTATION.lower()}-{parameters.TITLE}.svg')
+    if not os.path.exists(f'A3-{parameters.ORIENTATION.lower()}-{parameters.TITLE}.svg'):
+        shutil.copy2(f'../A3-{parameters.ORIENTATION.lower()}.svg', f'A3-{parameters.ORIENTATION.lower()}-{parameters.TITLE}.svg')
